+++ beetsplug/discogs.py
@@ -20,8 +20,8 @@ from __future__ import division, absolute_import, print_function
 
 import beets.ui
 from beets import config
-from beets.autotag.hooks import AlbumInfo, TrackInfo
-from beets.plugins import MetadataSourcePlugin, BeetsPlugin, get_distance
+from beets.autotag.hooks import AlbumInfo, TrackInfo, get_distance
+from beets.plugins import MetadataSourcePlugin, BeetsPlugin
 import confuse
 from discogs_client import Release, Master, Client
 from discogs_client.exceptions import DiscogsAPIError
